'global' chunk label (mirrors shiny global.R)
Shiny-based flexdashboards with multiple users can currently specify expensive one-time startup operations using an external global.R file. This PR enables this code to be directly incorporated into the Rmd file using a special chunk with the label 'global'.

The 'global' chunk's evaluation is customized using a kntir evaluate hook as follows:

1) It's evaluated in the global environment;

2) The actual code evaluated it tracked to ensure that the code is never evaluated more than once (i.e. subsequent users in a multi-user application benefit from the one time initialization).

The motivation for this PR is largely to support documenting best practices for performance in single-file examples (we have many of them now their code can be inspected via the prominent "Source Code" button in the top-right of the dashboard).